# api/openapi.yaml
openapi: 3.1.0
info:
  title: Quill Cloud API
  version: 2.6.0
  description: REST API for durable tasks, queues and schedules.
servers:
  - url: https://api.quillstack.example/v1
security:
  - bearerAuth: []
paths:
  /tasks:
    get:
      summary: List tasks
      responses:
        '200': { description: Task list }
    post:
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [queue, payload]
              properties:
                queue: { type: string }
                payload: { type: object, additionalProperties: { type: string } }
                priority: { type: string, enum: [low, normal, high], default: normal, description: Task priority; higher-priority tasks are dequeued first. }
                retries: { type: integer, minimum: 0, default: 5, description: Retry count; defaults to 5 when omitted. }
                ttl_seconds: { type: integer, minimum: 1, description: Expire a task if it remains queued for this many seconds; omitted means no expiry. }
      responses:
        '201': { description: Task created }
  /tasks/{id}/cancel:
    post:
      summary: Cancel a task
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: Task cancelled }
        '404': { description: Task not found }
        '409': { description: Task is already finished or cancelled }
  /queues:
    get:
      summary: List queues
      responses:
        '200':
          description: Queue list; max_concurrency is an integer or "unlimited" when no limit is set.
    post:
      summary: Create a queue
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, concurrency]
              properties:
                name: { type: string }
                concurrency: { type: integer, minimum: 1 }
                max_concurrency: { type: integer, minimum: 1, description: Maximum concurrently running tasks from this queue; omitted means unlimited. }
                retries: { type: integer, minimum: 0, default: 5, description: Retry count; defaults to 5 when omitted. }
                maxAttempts: { type: integer, minimum: 0, deprecated: true, description: Legacy alias for retries. }
      responses: { '201': { description: Queue created } }
  /queues/{name}/pause:
    post:
      summary: Pause a queue
      description: New tasks remain accepted, but workers cannot claim queued tasks until the queue is resumed.
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '200': { description: Queue paused }
        '404': { description: Queue not found }
  /queues/{name}/resume:
    post:
      summary: Resume a queue
      parameters:
        - in: path
          name: name
          required: true
          schema: { type: string }
      responses:
        '200': { description: Queue resumed }
        '404': { description: Queue not found }
  /schedules:
    get: { summary: List schedules, responses: { '200': { description: Schedule list } } }
    post: { summary: Create a schedule, responses: { '201': { description: Schedule created } } }
  /webhooks:
    post: { summary: Register a webhook endpoint, responses: { '201': { description: Endpoint registered } } }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer